Protect: private APIs using guards

+++ src/modules/user/controllers/user.controller.ts
@@ -1,7 +1,11 @@
-import { Controller, Get, Post, Body, Param, Delete, Put } from '@nestjs/common';
+import { Controller, Get, Post, Body, Param, Delete, Put, UseGuards, Req, UnauthorizedException } from '@nestjs/common';
 import { UserService } from '../services/user.service';
 import { CreateUserDto } from '../dto/create-user.dto';
 import { UpdateUserDto } from '../dto/update-user.dto';
+import { JwtAuthGuard } from 'src/modules/common/jwt-guard/jwt-auth.guard';
+import { User } from 'src/modules/database/entities';
+import { get } from 'lodash';
+
 
 @Controller('api/user')
 export class UserController {
@@ -12,23 +16,40 @@ export class UserController {
     return this.userService.createUserHandler(createUserDto);
   }
 
-  @Get()
-  findAll() {
-    return this.userService.findAll();
-  }
-
+  @UseGuards(JwtAuthGuard)
   @Get('/:id')
-  findOne(@Param('id') id: string) {
+  findOne(
+    @Param('id') id: string,
+    @Req() req: Request
+    ) {
+    if(get(req, "user.id") != id) {
+      throw new UnauthorizedException("You cannot perform this operation for given user")
+    }
     return this.userService.findOne(+id);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Put('/:id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  update(
+    @Param('id') id: string,
+    @Body() updateUserDto: UpdateUserDto,
+    @Req() req: Request
+  ) {
+    if(get(req, "user.id") != id) {
+      throw new UnauthorizedException("You cannot perform this operation for given user")
+    }
     return this.userService.update(+id, updateUserDto);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Delete('/:id')
-  remove(@Param('id') id: string) {
+  remove(
+    @Param('id') id: string,
+    @Req() req: Request
+    ) {
+    if(get(req, "user.id") != id) {
+      throw new UnauthorizedException("You cannot perform this operation for given user")
+    }
     return this.userService.remove(+id);
   }
 }
